+++ src/main/scala/com/atomist/rug/test/gherkin/ExecutableFeatureFactory.scala
@@ -8,9 +8,5 @@ import com.atomist.source.ArtifactSource
   */
 trait ExecutableFeatureFactory {
 
-  def executableFeatureFor(f: FeatureDefinition,
-                           definitions: Definitions,
-                           rugAs: ArtifactSource,
-                           rugs: Option[Rugs],
-                           listeners: Seq[GherkinExecutionListener]): AbstractExecutableFeature[_]
+  def executableFeatureFor(f: FeatureDefinition, definitions: Definitions, rugAs: ArtifactSource, rugs: Option[Rugs], listeners: Seq[GherkinExecutionListener], config: GherkinRunnerConfig): AbstractExecutableFeature[_]
 }