+++ src/main/scala/com/atomist/rug/test/gherkin/GherkinRunner.scala
@@ -1,21 +1,29 @@
 package com.atomist.rug.test.gherkin
 
-import com.atomist.graph.GraphNode
 import com.atomist.project.archive.Rugs
 import com.atomist.rug.runtime.js.JavaScriptContext
-import com.atomist.rug.test.gherkin.project.ProjectManipulationFeature
 import com.typesafe.scalalogging.LazyLogging
-import gherkin.ast.{ScenarioDefinition, Step}
+
+/**
+  * Config to use when running tests. Tokens etc.
+  *
+  * @param oAuthToken token to use for Git cloning
+  */
+case class GherkinRunnerConfig(oAuthToken: Option[String] = None)
 
 /**
   * Combine Gherkin DSL BDD definitions with JavaScript backing code
   * and provide the ability to execute the tests
   *
   * @param jsc       JavaScript backed by a Rug archive
-  * @param rugs      other Rugs from the achive
+  * @param rugs      other Rugs from the archive
   * @param listeners optional execution listeners
+  * @param config    config to use for test running
   */
-class GherkinRunner(jsc: JavaScriptContext, rugs: Option[Rugs] = None, listeners: Seq[GherkinExecutionListener] = Nil)
+class GherkinRunner(jsc: JavaScriptContext,
+                    rugs: Option[Rugs] = None,
+                    listeners: Seq[GherkinExecutionListener] = Nil,
+                    config: GherkinRunnerConfig = GherkinRunnerConfig())
   extends LazyLogging {
 
   import GherkinRunner._
@@ -38,16 +46,16 @@ class GherkinRunner(jsc: JavaScriptContext, rugs: Option[Rugs] = None, listeners
   val features: Seq[FeatureDefinition] = GherkinReader.findFeatures(jsc.rugAs)
 
   private val executableFeatures = features.map(f =>
-    featureFactory.executableFeatureFor(f, definitions, jsc.rugAs, rugs, listeners))
+    featureFactory.executableFeatureFor(f, definitions, jsc.rugAs, rugs, listeners, config))
 
   /**
     * Execute all the tests in this archive that pass provided filter
     *
     * @param filter callback to filter features that should execute
     */
-  def execute(filter: (FeatureDefinition) => Boolean = (fd) => true): ArchiveTestResult = {
+  def execute(filter: (FeatureDefinition) => Boolean = _ => true): ArchiveTestResult = {
     logger.info(s"Execute on $this")
-    ArchiveTestResult(executableFeatures.filter(pmf => filter.apply(pmf.definition)).map(ef => jsc.withEnhancedExceptions {
+    ArchiveTestResult(executableFeatures.filter(pmf => filter(pmf.definition)).map(ef => jsc.withEnhancedExceptions {
       listeners.foreach(_.featureStarting(ef.definition))
       val result = ef.execute()
       listeners.foreach(_.featureCompleted(ef.definition, result))