+++ src/test/resources/com/atomist/rug/test/gherkin/project/RealEditSteps.ts
@@ -1,0 +1,7 @@
+
+// We need this to ensure that well-known steps are imported
+import { Given, When, Then, ProjectScenarioWorld } from "@atomist/rug/test/project/Core";
+
+Given("a visionary leader", p => {
+    //
+});